Changed Graphbuilder Node tracking, instead using name using Path
as the path of the files are absoulte, there is no way we can have duplicated paths, before creating a node we verify if the path has already a Node ID if not, we create one, if yes we recycle it.

Also added different shapes for class, lvlib, ctl, vis.